degiro integration : mvc split
