+++ gamestonk_terminal/portfolio/README.md
@@ -234,7 +234,12 @@ Displays current holdings to the console.
 
 ### lastnews <a name="lastnews"></a>
 ````
-usage: lastnews
+usage: lastnews [-l LIMIT] [-h]
+
+optional arguments:
+  -l LIMIT, --limit LIMIT
+                        Number of news to display.
+  -h, --help            show this help message
 ````
 
 ### login <a name="login"></a>
