+++ gamestonk_terminal/portfolio/port_controller.py
@@ -12,7 +12,7 @@ from gamestonk_terminal.portfolio import (
     rh_api,
     alp_api,
     ally_api,
-    degiro_api,
+    degiro_controller,
 )
 from gamestonk_terminal.portfolio.portfolio_helpers import (
     merge_portfolios,
@@ -173,7 +173,7 @@ class PortfolioController:
     def call_degiro(self, _):
         """ "Process degiro command."""
 
-        return degiro_api.menu()
+        return degiro_controller.menu()
 
     def call_hold(self, _):
         holdings = pd.DataFrame(
